Create remote file download
In this code, the PHP section handles the file downloading functionality. I have done this for knowledgemandi.com website.It checks if a file URL is provided via the $_POST superglobal and saves the file locally using file_put_contents() and file_get_contents() functions. The HTML section creates a form with an input field (file_url) where the user can enter the URL of the file to be downloaded. Upon submitting the form, the PHP code is executed, and appropriate messages are displayed based on the outcome of the download operation.

Make sure to update the 'path/to/save/' value in the $savePath variable to the desired directory where you want to save the downloaded file.